refactor auth flow & modify button variant
