--- frontend/app/(auth)/create-group.tsx
@@ -1,14 +1,0 @@
-import CreateGroupComponent from "@/design-system/components/groups/create-group";
-import { Keyboard, SafeAreaView, TouchableWithoutFeedback } from "react-native";
-
-const CreateGroup = () => {
-  return (
-    <TouchableWithoutFeedback onPress={Keyboard.dismiss}>
-      <SafeAreaView className="flex-1 mt-[25%]">
-        <CreateGroupComponent nextPageNavigate="/(auth)/invite-link" />
-      </SafeAreaView>
-    </TouchableWithoutFeedback>
-  );
-};
-
-export default CreateGroup;
