+++ frontend/app/(auth)/group/invite.tsx
@@ -1,0 +1,12 @@
+import InviteLinkComponent from "@/design-system/components/groups/invite-group";
+import { SafeAreaView } from "react-native";
+
+const InviteLink = () => {
+  return (
+    <SafeAreaView className="flex-1 mt-[25%]">
+      <InviteLinkComponent />
+    </SafeAreaView>
+  );
+};
+
+export default InviteLink;
