--- frontend/app/(auth)/invite-link.tsx
@@ -1,12 +1,0 @@
-import InviteLinkComponent from "@/design-system/components/groups/invite-group";
-import { SafeAreaView } from "react-native";
-
-const InviteLink = () => {
-  return (
-    <SafeAreaView className="flex-1 mt-[25%]">
-      <InviteLinkComponent />
-    </SafeAreaView>
-  );
-};
-
-export default InviteLink;
