+++ frontend/design-system/base/config/color.ts
@@ -1,13 +1,12 @@
 export const ColorPalette = {
   pearl: "#F1F1F1",
-  blush: "#FFCADA",
   honey: "#F5AE58",
   slate: "#524e68",
   ink: "#100B2D",
   error: "#FF0033",
   success: "#2ECC71",
   warning: "#F39C12",
-  darkGray: "#757575",
+  darkGray: "#5B4E4C",
   gray: "#CFCFCF",
   silver: "#DEDEDE",
 };
