+++ frontend/design-system/components/posts/carousel.tsx
@@ -127,7 +127,7 @@ const ImageCarousel: React.FC<CarouselProps> = ({ data, initialPage = 0, like, s
         {containerWidth > 0 && (
           <>
             <Box position="absolute" zIndex={10} right={0} bottom={0} padding="m">
-              <Heart variant="blush" onLike={handleLike} like={like} />
+              <Heart onLike={handleLike} like={like} />
             </Box>
 
             {showFlyingHeart && (
