+++ frontend/package.json
@@ -33,12 +33,12 @@
     "@supabase/supabase-js": "^2.47.6",
     "@tanstack/react-query": "^5.62.0",
     "class-variance-authority": "^0.7.1",
-    "expo": "52.0.36",
+    "expo": "52.0.38",
     "expo-app-loading": "^2.1.1",
     "expo-av": "^15.0.2",
     "expo-blur": "14.0.3",
-    "expo-camera": "16.0.17",
-    "expo-constants": "17.0.7",
+    "expo-camera": "16.0.18",
+    "expo-constants": "17.0.8",
     "expo-device": "^7.0.2",
     "expo-font": "13.0.4",
     "expo-haptics": "~14.0.1",
@@ -49,8 +49,8 @@
     "expo-local-authentication": "~15.0.2",
     "expo-localization": "^16.0.1",
     "expo-location": "18.0.7",
-    "expo-notifications": "0.29.13",
-    "expo-router": "4.0.17",
+    "expo-notifications": "0.29.14",
+    "expo-router": "4.0.18",
     "expo-secure-store": "~14.0.1",
     "expo-splash-screen": "0.29.22",
     "expo-status-bar": "~2.0.1",
@@ -91,7 +91,7 @@
     "eslint": "^8.57.0",
     "eslint-config-expo": "^8.0.1",
     "jest": "^29.2.1",
-    "jest-expo": "52.0.5",
+    "jest-expo": "52.0.6",
     "react-native-svg-transformer": "^1.5.0",
     "react-test-renderer": "18.3.1",
     "tailwindcss": "3.3.2",
